+++ truck-meshalgo/tests/filters/normal_filter.rs
@@ -107,7 +107,7 @@ fn add_naive_normals_test() {
 fn add_smooth_normals() {
     let mut sphere = common::shapes::sphere(Point3::origin(), 1.0, 20, 10);
     sphere
-        .put_together_same_attrs()
+        .put_together_same_attrs(TOLERANCE)
         .remove_degenerate_faces()
         .remove_unused_attrs();