+++ truck-meshalgo/tests/tessellation/triangulation.rs
@@ -20,7 +20,7 @@ fn solid_is_closed() {
     for (i, json) in read_jsons().into_iter().enumerate() {
         let solid: Solid = serde_json::from_reader(json.as_slice()).unwrap();
         let mut poly = solid.triangulation(0.01).to_polygon();
-        poly.put_together_same_attrs()
+        poly.put_together_same_attrs(TOLERANCE * 2.0)
             .remove_degenerate_faces()
             .remove_unused_attrs();
         assert_eq!(
@@ -36,7 +36,7 @@ fn csolid_is_closed() {
     for (i, json) in read_jsons().into_iter().enumerate() {
         let solid: CompressedSolid = serde_json::from_reader(json.as_slice()).unwrap();
         let mut poly = solid.triangulation(0.01).to_polygon();
-        poly.put_together_same_attrs()
+        poly.put_together_same_attrs(TOLERANCE * 2.0)
             .remove_degenerate_faces()
             .remove_unused_attrs();
         assert_eq!(
@@ -51,7 +51,7 @@ fn csolid_is_closed() {
 fn compare_occt_mesh() {
     let jsons = read_jsons();
     let solid: Solid = serde_json::from_slice(jsons[2].as_slice()).unwrap();
-    let res = solid.triangulation(0.01).to_polygon();
+    let res = solid.triangulation(0.005).to_polygon();
     let path = concat!(dir!(), "../obj/by_occt.obj");
     let ans = obj::read(std::fs::read(path).unwrap().as_slice()).unwrap();
     assert!(res.is_clung_to_by(ans.positions(), 0.05));
@@ -62,7 +62,7 @@ fn compare_occt_mesh() {
 fn compare_occt_mesh_csolid() {
     let jsons = read_jsons();
     let solid: CompressedSolid = serde_json::from_slice(jsons[2].as_slice()).unwrap();
-    let res = solid.triangulation(0.01).to_polygon();
+    let res = solid.triangulation(0.005).to_polygon();
     let path = concat!(dir!(), "../obj/by_occt.obj");
     let ans = obj::read(std::fs::read(path).unwrap().as_slice()).unwrap();
     assert!(res.is_clung_to_by(ans.positions(), 0.05));
@@ -116,7 +116,7 @@ fn special_cylinder_model() -> Shell {
 fn special_cylinder() {
     let shell = special_cylinder_model();
     let mut mesh = shell.triangulation(0.01).to_polygon();
-    mesh.put_together_same_attrs()
+    mesh.put_together_same_attrs(TOLERANCE)
         .remove_degenerate_faces()
         .remove_unused_attrs();
     assert_eq!(mesh.shell_condition(), ShellCondition::Closed);
@@ -126,7 +126,7 @@ fn special_cylinder() {
 fn special_cylinder_csolid() {
     let shell = special_cylinder_model().compress();
     let mut mesh = shell.triangulation(0.01).to_polygon();
-    mesh.put_together_same_attrs()
+    mesh.put_together_same_attrs(TOLERANCE)
         .remove_degenerate_faces()
         .remove_unused_attrs();
     assert_eq!(mesh.shell_condition(), ShellCondition::Closed);