+++ truck-rendimpl/examples/simple-obj-viewer.rs
@@ -113,7 +113,7 @@ impl MyApp {
         reader: R,
     ) -> (PolygonInstance, WireFrameInstance) {
         let mut mesh = obj::read(reader).unwrap();
-        mesh.put_together_same_attrs()
+        mesh.put_together_same_attrs(TOLERANCE * 2.0)
             .add_smooth_normals(0.5, false);
         let bdd_box = mesh.bounding_box();
         let (size, center) = (bdd_box.size(), bdd_box.center());